# backend/docs/api-spec.yaml
openapi: 3.0.0
info:
  title: Felparkering API
  description: Rest API för att hantera felparkerade fordon
  contact: {}
  version: '1.0'
servers:
  - url: http://localhost:8080
    variables: {}
paths:
  /login:
    post:
      tags:
        - Auth
      summary: Login user
      description: |-
        ### POST /login

        Loggar in användare med e-post och lösenord.  
        Returnerar en JWT-token som används för autentisering.
      operationId: Loginuser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: admin@example.com
                password:
                  type: string
                  example: abc123
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                userDetails:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
                message: User logged in successfully
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorMissingCredentials'
                  - example:
                      error: Missing credentials
              example:
                error: Missing credentials
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorInvalidCredentials'
                  - example:
                      error: Invalid credentials
              example:
                error: Invalid credentials
        '404':
          description: Not Found
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorNotFound'
                  - example:
                      error: User not found
              example:
                error: User not found
      deprecated: false
      security: []
  /register:
    post:
      tags:
        - Auth
      summary: Register new user
      description: |-
        ### POST /register

        Registrerar en användare (CUSTOMER) med e-post och lösenord.
      operationId: Registernewuser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: admin@example.com
                password:
                  type: string
                  example: abc123
                confpassword:
                  type: string
                  example: abc123
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              example:
                  user: 
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
                  message: User registered successfully
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorMissingCredentials'
                  - $ref: '#/components/schemas/ErrorPasswordsDontMatch'
              examples:
                missingCredentials: 
                  summary: Missing credentials
                  value:
                    error: Missing credentials
                passwordMismatch:
                  summary: Password and confirmation does not match
                  value:
                    error: Password and confirmation does not match
        '409':
          description: Conflict
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ErrorUserConflict'
                  - example:
                      message: 
                      error: User already exists
              example:
                error: User already exists
      deprecated: false
...
components:
  schemas:
    LoginResponse:
      title: LoginResponse
      required:
        - user
        - message
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserDetailDto'
        message:
          type: string
      example:
        user:
          token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
        message: User logged in successfully
    ErrorMissingCredentials:
      title: ErrorMissingCredentials
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: Missing credentials
    ErrorInvalidCredentials:
      title: ErrorInvalidCredentials
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: Invalid credentials
    ErrorNotFound:
      title: ErrorNotFound
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: User not found
    ErrorUserConflict:
      title: ErrorUserConflict
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: User already exists
    ErrorPasswordsDontMatch:
      title: ErrorPasswordsDontMatch
      required:
        - message
      type: object
      properties:
        error:
          type: string
      example:
        error: Password and confirmation does not match
    RegisterResponse:
      title: RegisterResponse
      required:
        - token
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserDetailDto'
        message:
          type: string
      example:
        user:
          token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
        message: User registered successfully
...
    UserDetailDto:
      title: UserDetailDto
      required:
        - token
      type: object
      properties:
        token:
          type: string
      example:
          token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...